+++ widgets/TaskPicker.php
@@ -177,12 +177,9 @@ class TaskPicker extends \yii\base\Widget
 
         $cfg = ($cfg == null) ? $defaultCfg : array_merge($defaultCfg, $cfg);
 
-        //If no initial query is given we use getFriends if friendship module is enabled otherwise all tasks
-        if (!isset($cfg['query'])) {
-            $cfg['query'] = (Yii::$app->getModule('friendship')->settings->get('enable'))
-                    ? Yii::$app->task->getIdentity()->getFriends()
-                    : UserFilter::find();
-        }
+        $cfg['query'] ??= (Yii::$app->getModule('friendship')->settings->get('enable'))
+                ? Yii::$app->task->getIdentity()->getFriends()
+                : UserFilter::find();
 
         //Filter the initial query and disable task without the given permission
         $task = UserFilter::filter($cfg['query'], $cfg['keyword'], $cfg['maxResult'], null, $cfg['active']);