+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 --------------------
 - Fix #329: Make action elements focusable in task list via keyboard
 - Fix #328: Fix keyboard navigation (Tab/arrow keys) between tabs in the task creation/edit modal form
+- Enh: Automated code refactoring for HumHub 1.18.0-beta.6 using Rector
 
 1.9.6 (July 8, 2026)
 --------------------